pyproject.toml
[project]
name = "chip-selection-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
]
